+++ pytorch/src/gen/java/org/bytedeco/pytorch/DictKeyEqualTo.java
@@ -16,23 +16,23 @@ import static org.bytedeco.openblas.global.openblas.*;
 import static org.bytedeco.pytorch.global.torch.*;
 
 
-@Name("c10::detail::getTypePtr_<at::IValue>") @Properties(inherit = org.bytedeco.pytorch.presets.torch.class)
-public class getTypePtr_ extends Pointer {
+@Namespace("c10::detail") @Properties(inherit = org.bytedeco.pytorch.presets.torch.class)
+public class DictKeyEqualTo extends Pointer {
     static { Loader.load(); }
     /** Default native constructor. */
-    public getTypePtr_() { super((Pointer)null); allocate(); }
+    public DictKeyEqualTo() { super((Pointer)null); allocate(); }
     /** Native array allocator. Access with {@link Pointer#position(long)}. */
-    public getTypePtr_(long size) { super((Pointer)null); allocateArray(size); }
+    public DictKeyEqualTo(long size) { super((Pointer)null); allocateArray(size); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
-    public getTypePtr_(Pointer p) { super(p); }
+    public DictKeyEqualTo(Pointer p) { super(p); }
     private native void allocate();
     private native void allocateArray(long size);
-    @Override public getTypePtr_ position(long position) {
-        return (getTypePtr_)super.position(position);
+    @Override public DictKeyEqualTo position(long position) {
+        return (DictKeyEqualTo)super.position(position);
     }
-    @Override public getTypePtr_ getPointer(long i) {
-        return new getTypePtr_((Pointer)this).offsetAddress(i);
+    @Override public DictKeyEqualTo getPointer(long i) {
+        return new DictKeyEqualTo((Pointer)this).offsetAddress(i);
     }
 
-  public static native @ByVal IValue call();
+  public native @Cast("bool") @Name("operator ()") boolean apply(@Const @ByRef IValue lhs, @Const @ByRef IValue rhs);
 }